requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
matplotlib>=3.7
numpy>=1.24
playwright>=1.40
Pillow>=10.0
deep-translator>=1.11
sentence-transformers>=3.0